+++ src/main/java/com/company/springmvcweb/data/LightItem.java
@@ -12,13 +12,13 @@ import javax.persistence.*;
 
 public class LightItem extends Item{
 //@Column(name="power")
-    private int power;
+    private Float power;
 
-    public LightItem(int power) {
+    public LightItem(Float power) {
         this.power = power;
     }
 
-    public LightItem(int id, String name, Integer price, String pic, boolean isActive, Location location, int totalCount, int power) {
+    public LightItem(int id, String name, Float price, String pic, boolean isActive, Location location, int totalCount, Float power) {
         super(id, name, price, pic, isActive, location, totalCount);
         this.power = power;
     }
