+++ src/main/java/com/company/springmvcweb/data/TrussItem.java
@@ -10,9 +10,9 @@ import javax.persistence.Entity;
 @DiscriminatorValue(value=Category.CatValues.TRUSS)
 public class TrussItem extends Item{
 //    @Column(name="length")
-    private int length;
+    private Integer length;
 
-    public TrussItem(int id, String name, Integer price, String pic, boolean isActive, Location location, int totalCount, int length) {
+    public TrussItem(int id, String name, Float price, String pic, boolean isActive, Location location, int totalCount, Integer length) {
         super(id, name, price, pic, isActive, location, totalCount);
         this.length = length;
     }
