+++ src/main/java/com/company/springmvcweb/data/VideoItem.java
@@ -10,7 +10,7 @@ import javax.persistence.Entity;
 @DiscriminatorValue(value=Category.CatValues.VIDEO)
 public class VideoItem extends Item{
 
-    public VideoItem(int id, String name, Integer price, String pic, boolean isActive, Location location, int totalCount) {
+    public VideoItem(int id, String name, Float price, String pic, boolean isActive, Location location, int totalCount) {
         super(id, name, price, pic, isActive, location, totalCount);
     }
